+++ __tests__/data_structures.ts
@@ -1,14 +1,5 @@
-import { CapabilitiesRequest } from '../proto/world-engine.pb';
-import { Capabilities } from '../src/common/data_structures';
+import { HistoryItem } from '../src/components/history';
 import { InworldPacket } from '../src/entities/inworld_packet.entity';
-
-export interface ExtendedCapabilities extends Capabilities {
-  regenerateResponse: boolean;
-}
-
-export interface ExtendedCapabilitiesRequest extends CapabilitiesRequest {
-  regenerateResponse: boolean;
-}
 
 export interface RegenerateResponse {
   interactionId?: string;
@@ -21,3 +12,5 @@ export interface MutationEvent {
 export interface ExtendedInworldPacket extends InworldPacket {
   mutation: MutationEvent;
 }
+
+export type ExtendedHistoryItem = HistoryItem & { mutation: MutationEvent };